Added api for cancel order and fixed other api issue and model.

+++ Core/core/src/main/java/com/yilinker/core/model/TransactionProduct.java
@@ -214,11 +214,11 @@ public class TransactionProduct {
         this.unitPrice = unitPrice;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
